pyproject.toml
[project]
name = "validate-git-blame-ignore-revs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eric Berquist", email = "eric.berquist@gmail.com" }
]
requires-python = ">=3.9"
dependencies = []

[project.scripts]
validate-git-blame-ignore-revs = "validate_git_blame_ignore_revs.__main__:main"

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "dulwich>=0.24.1",
    "pygit2>=1.15.1",
    "pytest-cov>=6.2.1",
]

[tool.mypy]
strict = true

[tool.ruff]
line-length = 100
